HIVE-1719 [jira] Move RegexSerDe out of hive-contrib and over to hive-serde (Shreepadma Venugopalan via Carl Steinbach)
Summary:
Regex Serde Changes

RegexSerDe is as much a part of the standard Hive distribution as the other SerDes
currently in hive-serde. I think we should move it over to the hive-serde module so that
users don't have to go to the added effort of manually registering the contrib jar before
using it.

Test Plan: EMPTY

Reviewers: JIRA, cwsteinbach

Reviewed By: cwsteinbach

Differential Revision: https://reviews.facebook.net/D3249

git-svn-id: https://svn.apache.org/repos/asf/hive/trunk@1340256 13f79535-47bb-0310-9956-ffa450edef68