Track tables and keys used in join and group by operators (Sambavi Muthukrishnan via Carl Steinbach)
Summary:
The idea here is to walk the operator tree and identify join and group by operators
whose keys can be mapped back directly to tables and columns on those tables. This
information can be used to determine the characteristics of a workload and which
tables are accessed most often using the same keys for these operators that can 
benefit from bucketing.

Test Plan: Added new test cases in TestCliDriver. Ran all existing unit tests and they passed.

Reviewers: njain, kevinwilfong, JIRA, cwsteinbach

Reviewed By: cwsteinbach

CC: cwsteinbach

Differential Revision: https://reviews.facebook.net/D5613

git-svn-id: https://svn.apache.org/repos/asf/hive/trunk@1393935 13f79535-47bb-0310-9956-ffa450edef68